pyproject.toml
[project]
name = "hackyeah-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.8.0",
    "fastapi>=0.118.0",
    "google-api-python-client>=2.184.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-pubsub>=2.31.1",
    "google-genai>=1.41.0",
    "numpy>=2.3.3",
    "oauth2client>=4.1.3",
    "opencv-python>=4.11.0.86",
    "pip>=25.2",
    "pytz>=2025.2",
    "requests>=2.32.5",
]
